refactor: reorganize folders, move API calls to repository, cleanup client usage
